+++ CHANGES.md
@@ -1,5 +1,13 @@
 # PlayCanvas Engine Changes
 
+### v0.168.6
+* [FIX] Destroy AudioManager when application is destroyed.
+* [FIX] Better handling of application destruction while resources are being preloaded.
+
+### v0.168.5
+* [FIX] Added fix for instancing.
+* Added pc.Application#destroy.
+
 ### v0.168.4
 * [FIX] Materials cubemap property works again.